Devi Durga and rainfall
The Hindu goddess Durga is said to come on a carrier and leave by one each year and as those dates according to lunar calendar fall on any of the seven days of a week in solar calendar, the carrier is predicted and it is said to be associated with the amount of harvest or rainfall in that particular year. Let's see how relevant is that.
The data about the carrier of the goddess is taken from the website of DrikPanchang( https://www.drikpanchang.com ) using a self-made simple chrome extension for years 1900 to 2016.
The data on rainfall is available on kaggle( https://www.kaggle.com/rajanand/rainfall-in-india/data ) from 1901-2015.